requirements.txt
pyside6
psycopg2
django==4.2.10
djangorestframework==3.14.0
